feat: use highest priority request match error
When a request fails to match against multiple possible rules, a
comparison function ranks the priority of errors so that the single most
relavant error is returned in the end.

For example, if we get an authentication and authorisation error, the
authorisation error is used, as it implies authentication succeeded at
least once.