feat: test scripts/config.ts and make env output the default
Markdown output doesn't really make sense as the default in retrospect.
The typical use case is probably to write a JSON config, then use
scripts/config.ts to validate it and generate the environment variables
needed for Deno Deploy.

While the markdown output gives you that info too, it's a lot of noise
that's probably not helpful.