feat: individual field overrides in scripts/config.ts
We now allow a config loaded from a JSON file to have individual fields
selectively overridden by values from the environment (or command line
arguments). This allows for secrets in environment variables to merge
over a config file.

The use case I have for this is a CI run with an envar set from a CI
secret, with the main config file committed in the repo.